+++ src/test/java/org/stellar/sdk/MuxedAccountTest.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.assertThrows;
 
 import java.math.BigInteger;
 import org.junit.Test;
-import org.stellar.sdk.exception.FormatException;
+import org.stellar.sdk.exception.StrKeyException;
 import org.stellar.sdk.xdr.CryptoKeyType;
 import org.stellar.sdk.xdr.Uint256;
 import org.stellar.sdk.xdr.Uint64;
@@ -75,6 +75,6 @@ public class MuxedAccountTest {
   public void testFromAccountInvalidAccountRaise() {
     String invalidAccount =
         "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA";
-    assertThrows(FormatException.class, () -> new MuxedAccount(invalidAccount));
+    assertThrows(StrKeyException.class, () -> new MuxedAccount(invalidAccount));
   }
 }