+++ CHANGELOG.md
@@ -75,6 +75,7 @@ As this project is pre 1.0, breaking changes may happen for minor version bumps.
 - refactor!: `TransactionPreconditions#TransactionPreconditions(LedgerBounds, Long, BigInteger, long, List, TimeBounds)` has been removed, use `TransactionPreconditions#TransactionPreconditions(TimeBounds, LedgerBounds, Long, BigInteger, long, List)` instead.
 - refactor: Set the default value of `TransactionPreconditions.extraSigners` to `new ArrayList<>()`, it is not nullable.
 - refactor!: The `Federation` has been refactored, please use `Federation#resolveAddress(String)` and `Federation#resolveAccountId(String, String)` now.
+- refactor!: `FormatException` has been renamed to `StrKeyFormatException`. When encode or decode strkey fails, `StrKeyFormatException` will be thrown.
 
 ## 0.44.0
 ### Update